feat: Added required authorization to all endpoints that modify data

+++ src/app/api/survey/[surveyId]/question/[questionId]/route.ts
@@ -1,4 +1,4 @@
-import { BadRequest, NoContent, NotFound, Ok, RouteParams } from "@/lib/routeHelper";
+import { BadRequest, NoContent, NotFound, Ok, RequireAuthorization, RouteParams, Unauthorized } from "@/lib/routeHelper";
 import { QuestionUpdateDTOFromJSON } from "@/models/dto/questionUpdateDTO";
 import { DeleteQuestion, GetQuestion, UpdateQuestion } from "@/repository/questionRepository";
 import { RemoveQuestionFromOrder } from "@/repository/surveyRepository";
@@ -22,6 +22,10 @@ export async function GET(_: NextRequest, {params}: RouteParams<Params>) {
 
 export async function PUT(request: NextRequest, { params }: RouteParams<Params>) {
     try {
+        if(!RequireAuthorization(request)) {
+            return Unauthorized();
+        }
+
         const json = await request.json();
 
         const dto = QuestionUpdateDTOFromJSON(json);
@@ -39,7 +43,11 @@ export async function PUT(request: NextRequest, { params }: RouteParams<Params>)
 }
 
 
-export async function DELETE(_: NextRequest, { params }: RouteParams<Params>) {
+export async function DELETE(request: NextRequest, { params }: RouteParams<Params>) {
+    if(!RequireAuthorization(request)) {
+        return Unauthorized();
+    }
+
     await DeleteQuestion(params.surveyId, params.questionId);
 
     await RemoveQuestionFromOrder(params.surveyId, params.questionId);
