+++ src/app/api/survey/[surveyId]/route.ts
@@ -1,6 +1,7 @@
-import { BadRequest, NoContent, NotFound, Ok, RouteParams } from "@/lib/routeHelper";
+import { BadRequest, NoContent, NotFound, Ok, RequireAuthorization, RouteParams, Unauthorized } from "@/lib/routeHelper";
 import { SurveyUpdateDTOFromJSON } from "@/models/dto/surveyUpdateDTO";
 import { DeleteSurvey, GetSurvey, UpdateSurvey } from "@/repository/surveyRepository";
+import { request } from "http";
 import { NextRequest } from "next/server";
 
 interface Params {
@@ -22,6 +23,10 @@ export async function GET(_: NextRequest, { params }: RouteParams<Params>) {
 
 export async function PUT(request: NextRequest, { params }: RouteParams<Params>) {
     try {
+        if(!RequireAuthorization(request)) {
+            return Unauthorized();
+        }
+
         const json = await request.json();
 
         const dto = SurveyUpdateDTOFromJSON(json);
@@ -39,7 +44,11 @@ export async function PUT(request: NextRequest, { params }: RouteParams<Params>)
     }
 }
 
-export async function DELETE(_: NextRequest, { params }: RouteParams<Params>) {
+export async function DELETE(request: NextRequest, { params }: RouteParams<Params>) {
+    if(!RequireAuthorization(request)) {
+        return Unauthorized();
+    }
+
     await DeleteSurvey(params.surveyId)
 
     return NoContent();
